requirements.txt
Flask==2.3.2
Pillow==9.5.0
Flask-Cors==4.0.1